--- a/sktime/tests/test_switch.py
+++ b/sktime/tests/test_switch.py
@@ -10,7 +10,6 @@ from functools import lru_cache
 from inspect import getmro, isclass
 
 
-@lru_cache
 def run_test_for_class(cls):
     """Check if test should run for a class or function.
 
@@ -59,8 +58,15 @@ def run_test_for_class(cls):
     bool : True if class should be tested, False otherwise
         if cls was a list, is True iff True for at least one of the classes in the list
     """
-    if not isinstance(cls, list):
-        cls = [cls]
+    if isinstance(cls, list):
+        return all(_run_test_for_class(x) for x in cls)
+    else:
+        return _run_test_for_class(cls)
+
+
+@lru_cache
+def _run_test_for_class(cls):
+    """Check if test should run - cached with hashable cls."""
 
     from sktime.tests.test_all_estimators import ONLY_CHANGED_MODULES
     from sktime.utils.git_diff import get_packages_with_changed_specs, is_class_changed
@@ -113,7 +119,7 @@ def run_test_for_class(cls):
 
     # Condition 1:
     # if any of the required soft dependencies are not present, do not run the test
-    if not all(_required_deps_present(x) for x in cls):
+    if not _required_deps_present(cls):
         return False
     # otherwise, continue
 
@@ -125,16 +131,16 @@ def run_test_for_class(cls):
     # Condition 2:
     # any of the modules containing any of the classes in the list have changed
     # or any of the modules containing any parent classes in sktime have changed
-    cond2 = any(_is_class_changed_or_sktime_parents(x) for x in cls)
+    cond2 = _is_class_changed_or_sktime_parents(cls)
 
     # Condition 3:
     # if the object is an sktime BaseObject, and one of the test classes
     # covering the class have changed, then run the test
-    cond3 = any(_tests_covering_class_changed(x) for x in cls)
+    cond3 = _tests_covering_class_changed(cls)
 
     # Condition 4:
     # the package requirements for any dependency in pyproject.toml have changed
-    cond4 = any(_is_impacted_by_pyproject_change(x) for x in cls)
+    cond4 = _is_impacted_by_pyproject_change(cls)
 
     # run the test if and only if at least one of the conditions 2-4 are met
     return cond2 or cond3 or cond4