--- a/sktime/tests/tests/__init__.py
+++ b/sktime/tests/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the test utilities."""