--- a/sktime/utils/git_diff.py
+++ b/sktime/utils/git_diff.py
@@ -3,11 +3,13 @@
 __author__ = ["fkiraly"]
 __all__ = []
 
+from functools import lru_cache
 import importlib.util
 import inspect
 import subprocess
 
 
+@lru_cache
 def get_module_from_class(cls):
     """Get full parent module string from class.
 
@@ -24,6 +26,7 @@ def get_module_from_class(cls):
     return module.__name__ if module else None
 
 
+@lru_cache
 def get_path_from_module(module_str):
     r"""Get local path string from module string.
 
@@ -47,6 +50,7 @@ def get_path_from_module(module_str):
         raise ImportError(f"Error finding module '{module_str}'") from e
 
 
+@lru_cache
 def is_module_changed(module_str):
     """Check if a module has changed compared to the main branch.
 
@@ -64,6 +68,7 @@ def is_module_changed(module_str):
         return True
 
 
+@lru_cache
 def is_class_changed(cls):
     """Check if a class' parent module has changed compared to the main branch.
 
@@ -121,6 +126,7 @@ def get_changed_lines(file_path, only_indented=True):
         return []
 
 
+@lru_cache
 def get_packages_with_changed_specs():
     """Get packages with changed or added specs.