--- a/sktime/utils/git_diff.py
+++ b/sktime/utils/git_diff.py
@@ -4,6 +4,7 @@ __author__ = ["fkiraly"]
 __all__ = []
 
 from functools import lru_cache
+
 import importlib.util
 import inspect
 import subprocess