Towards https://github.com/sktime/sktime/issues/6344.

Speeds up test collection by `lru_cache`-ing all test switch utilities called from `run_test_for_class`.

This gets test collection time for the classification module to 5 seconds, compared to historical test times:

* 0.22.0 - 12 sec for 301 tests
* 0.24.0 - 11 sec for 294 tests
    * in 0.24.2, parent classes were included in diff check
* 0.25.0 - 38 sec for 355 tests
* 0.26.0 - 38 sec for 355 tests
    * in 0.26.1, the condition on dep change in `pyproject` was added
* 0.27.0 - 56 sec for 363 tests
* 0.28.1 - 72 sec for 246 tests
* this PR: 5 sec for 246 tests

(timings from `python -m pytest sktime/classification -n0 --collect-only`)

For all of `sktime`, the test collection time is down to 40sec for me locally, from >20 minutes.